Changed data type of TotalPrice from int to double in CartController, CartDetail.cshtml, and _Layout.cshtml

+++ WebSellFlower/Controllers/CartController.cs
@@ -108,6 +108,6 @@ namespace WebSellFlower.Controllers
 
 	public class CartItem : TblProduct
 	{		
-		public int? TotalPrice => Quantity * ProdPrice;
+		public double? TotalPrice => Quantity * ProdPrice;
 	}
 }
